+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v1.1.5
+## 11/24/2021
+
+1. [](#new)
+    * Added method `ObjectController::checkAuthorizations()` to check if one of the actions is true
+2. [](#bugfix)
+    * Fixed regression when calling flex router with a path
+
 # v1.1.4
 ## 11/16/2021